+++ src/brushes/BaseBrush.ts
@@ -1,9 +1,14 @@
 import type { Canvas } from '../canvas/Canvas';
-import { TEvent } from '../EventTypeDefs';
+import { CanvasEvents, TEvent } from '../EventTypeDefs';
+import { Observable } from '../Observable';
 import type { Point } from '../Point';
 import { Shadow } from '../Shadow';
-import { FabricObject } from '../shapes/Object/FabricObject';
-import { multiplyTransformMatrices } from '../util/misc/matrix';
+import type { FabricObject } from '../shapes/Object/FabricObject';
+import { TCachedFabricObject } from '../shapes/Object/Object';
+import {
+  invertTransform,
+  multiplyTransformMatrices,
+} from '../util/misc/matrix';
 import { sendObjectToPlane } from '../util/misc/planeChange';
 
 export type TBrushEventData = TEvent & { pointer: Point };
@@ -11,7 +16,10 @@ export type TBrushEventData = TEvent & { pointer: Point };
 /**
  * @see {@link http://fabricjs.com/freedrawing|Freedrawing demo}
  */
-export abstract class BaseBrush<T extends FabricObject = FabricObject> {
+export abstract class BaseBrush<
+  T extends FabricObject = FabricObject,
+  EventSpec extends CanvasEvents = CanvasEvents
+> extends Observable<EventSpec> {
   /**
    * Color of a brush
    * @type String
@@ -64,55 +72,65 @@ export abstract class BaseBrush<T extends FabricObject = FabricObject> {
   strokeDashArray: number[] | null = null;
 
   /**
-   * When `true`, the free drawing is limited to the whiteboard size. Default to false.
-   * @type Boolean
-   * @default false
-   */
-
-  limitedToCanvasSize = false;
-
-  /**
    * Same as FabricObject `clipPath` property.
    * The clip path is positioned relative to the top left corner of the viewport.
    * The `absolutePositioned` property renders the clip path w/o viewport transform.
    */
   clipPath?: FabricObject;
 
+  /**
+   * Cursor value used during free drawing
+   * @type String
+   * @default crosshair
+   */
+  cursor: CSSStyleDeclaration['cursor'] = 'crosshair';
+
   declare readonly canvas: Canvas;
 
   active = false;
 
+  private _disposer?: () => void;
+
   constructor(canvas: Canvas) {
+    super();
     this.canvas = canvas;
+    const subscribers = this.subscribe();
+    this._disposer = () => {
+      subscribers.forEach((d) => d());
+      this._disposer = undefined;
+    };
+  }
+
+  /**
+   * This method wires the internal lifecycle to canvas events,
+   * making it very easy to change the hooks that the brush responds to.
+   * @returns an array of disposers
+   */
+  protected subscribe(): VoidFunction[] {
+    return [this.on('resize', () => this._setBrushStyles())];
+  }
+
+  protected unsubscribe() {
+    this._disposer && this._disposer();
   }
 
   protected abstract _render(ctx: CanvasRenderingContext2D): void;
 
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  onMouseDown(pointer: Point, ev: TBrushEventData) {
-    this.active = true;
-  }
-
-  abstract onMouseMove(pointer: Point, ev: TBrushEventData): void;
-
-  /**
-   * @returns true if brush should continue blocking interaction
-   */
-  abstract onMouseUp(ev: TBrushEventData): void;
-
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  onDoubleClick(pointer: Point) {
-    // noop
-  }
-
   protected abstract finalizeShape(): T | undefined;
+
+  protected start() {
+    this.active = true;
+    this.canvas.setCursor(this.cursor);
+  }
 
   /**
    * Sets brush styles
    * @private
    * @param {CanvasRenderingContext2D} ctx
    */
-  _setBrushStyles(ctx: CanvasRenderingContext2D = this.canvas.contextTop) {
+  protected _setBrushStyles(
+    ctx: CanvasRenderingContext2D = this.canvas.contextTop
+  ) {
     ctx.strokeStyle = this.color;
     ctx.lineWidth = this.width;
     ctx.lineCap = this.strokeLineCap;
@@ -175,11 +193,29 @@ export abstract class BaseBrush<T extends FabricObject = FabricObject> {
   }
 
   /**
-   * needed for `absolutePositioned` `clipPath`
    * @private
+   * @param {CanvasRenderingContext2D} ctx
+   * @param {FabricObject} clipPath
    */
-  calcTransformMatrix() {
-    return this.canvas.viewportTransform;
+  private drawClipPathOnCache(
+    ctx: CanvasRenderingContext2D,
+    clipPath: TCachedFabricObject
+  ) {
+    ctx.save();
+    ctx.globalCompositeOperation = clipPath.inverted
+      ? 'destination-out'
+      : 'destination-in';
+    if (clipPath.absolutePositioned) {
+      ctx.transform(...invertTransform(this.canvas.viewportTransform));
+    }
+    clipPath.transform(ctx);
+    ctx.scale(1 / clipPath.zoomX, 1 / clipPath.zoomY);
+    ctx.drawImage(
+      clipPath._cacheCanvas,
+      -clipPath.cacheTranslationX,
+      -clipPath.cacheTranslationY
+    );
+    ctx.restore();
   }
 
   /**
@@ -187,24 +223,21 @@ export abstract class BaseBrush<T extends FabricObject = FabricObject> {
    * @param {CanvasRenderingContext2D} ctx
    * @param {FabricObject} clipPath
    */
-  drawClipPathOnCache(ctx: CanvasRenderingContext2D, clipPath: FabricObject) {
-    // TODO: no proto calls
-    FabricObject.prototype.drawClipPathOnCache.call(this, ctx, clipPath);
-  }
-
-  /**
-   * @private
-   * @param {CanvasRenderingContext2D} ctx
-   * @param {FabricObject} clipPath
-   */
-  _drawClipPath(ctx: CanvasRenderingContext2D, clipPath?: FabricObject) {
+  protected _drawClipPath(
+    ctx: CanvasRenderingContext2D,
+    clipPath?: FabricObject
+  ) {
     if (!clipPath) {
       return;
     }
-    ctx.save();
-    // TODO: no proto calls
-    FabricObject.prototype._drawClipPath.call(this, ctx, clipPath);
-    ctx.restore();
+    // needed to setup a couple of variables
+    // path canvas gets overridden with this one.
+    // TODO find a better solution?
+    clipPath._set('canvas', this.canvas);
+    clipPath.shouldCache();
+    clipPath._transformDone = true;
+    clipPath.renderCache({ forClipping: true });
+    this.drawClipPathOnCache(ctx, clipPath as TCachedFabricObject);
   }
 
   /**
@@ -220,7 +253,7 @@ export abstract class BaseBrush<T extends FabricObject = FabricObject> {
       clipPath,
       undefined,
       this.clipPath.absolutePositioned
-        ? multiplyTransformMatrices(this.calcTransformMatrix(), t)
+        ? multiplyTransformMatrices(this.canvas.viewportTransform, t)
         : t
     );
     return clipPath;