+++ src/brushes/DrawPoly.ts
@@ -1,11 +1,16 @@
+import { TPointerEventInfo } from '../EventTypeDefs';
+import { TFabricEvent } from '../FabricEvent';
 import { Point } from '../Point';
 import { Polygon } from '../shapes/Polygon';
 import { Polyline } from '../shapes/Polyline';
-import { TBrushEventData } from './BaseBrush';
 import { DrawShapeBase } from './DrawShapeBase';
 
 export class DrawPoly extends DrawShapeBase<Polyline> {
   builder = Polygon;
+
+  protected subscribe() {
+    return [...super.subscribe(), this.on('mouse:dblclick', this.end)];
+  }
 
   private addPoint(pointer: Point) {
     this.shape!.points.push(pointer);
@@ -34,8 +39,9 @@ export class DrawPoly extends DrawShapeBase<Polyline> {
     return shape;
   }
 
-  onMouseDown(pointer: Point, ev: TBrushEventData) {
-    super.onMouseDown(pointer, ev);
+  down(ev: TFabricEvent<TPointerEventInfo>) {
+    super.down(ev);
+    const { pointer } = ev;
     if (this.shape) {
       this.addPoint(pointer);
     } else {
@@ -45,18 +51,22 @@ export class DrawPoly extends DrawShapeBase<Polyline> {
     }
   }
 
-  onMouseMove(pointer: Point) {
-    this.replacePoint(pointer);
+  move(ev: TFabricEvent<TPointerEventInfo>) {
+    super.move(ev);
+    this.replacePoint(ev.pointer);
   }
 
-  onMouseUp({ pointer }: { pointer: Point }) {
+  up(ev: TFabricEvent<TPointerEventInfo>) {
+    super.up(ev);
+    const { pointer } = ev;
     this.replacePoint(pointer);
     this.addPoint(pointer);
     return true;
   }
 
-  onDoubleClick(pointer: Point) {
-    this.shape && this.replacePoint(pointer);
+  end(ev: TFabricEvent<TPointerEventInfo>) {
+    ev.preventDefault();
+    this.shape && this.replacePoint(ev.pointer);
     this.finalize();
   }
 }