+++ test/unit/canvas_events.js
@@ -112,7 +112,7 @@
 
     //  window resize
     cacheAndAssertTransformEvent();
-    canvas._onResize();
+    canvas._onWindowResize();
     assertTransformEventCacheIsReset();
   });
 
@@ -337,22 +337,29 @@
     assert.equal(isClick, false, 'moving the pointer, the click is false');
   });
 
-  QUnit.test('setDimensions and active brush', function(assert) {
+  QUnit.test('setDimensions and active brush', async function (assert) {
+    var done = assert.async();
     var prepareFor = false;
     var rendered = false;
     var canvas = new fabric.Canvas(null, { width: 500, height: 500 });
-    var brush = new fabric.PencilBrush({ color: 'red', width: 4 });
+    var brush = new fabric.SimpleBrush(canvas);
     canvas.isDrawingMode = true;
     canvas.freeDrawingBrush = brush;
     canvas.isCurrentlyDrawing = () => true;
-    brush.render = function() { rendered = true; };
-    brush._setBrushStyles = function() { prepareFor = true; };
-    canvas.setDimensions({ width: 200, height: 200 });
-    canvas.renderAll();
+    brush.render = () => { rendered = true; };
+    brush._setBrushStyles = () => { prepareFor = true };
+    await new Promise(resolve => {
+      Promise.all([
+        new Promise(resolve => canvas.on('resize', resolve)),
+        new Promise(resolve => canvas.on('after:render', resolve)),
+      ]).then(resolve);
+      canvas.setDimensions({ width: 200, height: 200 });
+    });
     assert.equal(rendered, true, 'the brush called the render method');
     assert.equal(prepareFor, true, 'the brush called the _setBrushStyles method');
+    done();
   });
-
+  
   QUnit.test('mouse:up should return target and currentTarget', function(assert) {
     var e1 = { clientX: 30, clientY: 30, which: 1 };
     var e2 = { clientX: 100, clientY: 100, which: 1 };