+++ test/visual/freedraw.js
@@ -1,17 +1,22 @@
-  function setBrush(canvas, brush) {
-    canvas.isDrawingMode = true;
-    canvas.freeDrawingBrush = brush;
-  }
-  var options = { e: { pointerId: 1 } };
-  async function pointDrawer(points, brush, onComplete = false, onMove = undefined) {
-    const { canvas } = brush;
-    canvas.calcViewportBoundaries();
+function setBrush(canvas, brush) {
+  canvas.isDrawingMode = true;
+  canvas.freeDrawingBrush = brush;
+  canvas.calcViewportBoundaries();
+}
+function fireBrushEvent(brush, type, pointer) {
+  brush.fire(`mouse:${type}:before`, fabric.Event.init({
+    e: { pointerId: 1 },
+    pointer
+  }));
+}
+async function pointDrawer(points, brush, onComplete = false, onMove = undefined) {
+  const { canvas } = brush;
     setBrush(canvas, brush);
-    brush.onMouseDown(points[0], options);
+    fireBrushEvent(brush, 'down', points[0]);
     for (var i = 1; i < points.length; i++) {
       points[i].x = parseFloat(points[i].x);
       points[i].y = parseFloat(points[i].y);
-      brush.onMouseMove(points[i], options);
+      fireBrushEvent(brush, 'move', points[i]);
       onMove && onMove(points[i], i, points);
     }
     if (onComplete) {
@@ -20,13 +25,13 @@
           typeof onComplete === 'function' ? onComplete(canvas, result) : canvas.add(result);
           resolve()
         });
-        brush.onMouseUp(options);
+        fireBrushEvent(brush, 'up', points[points.length - 1]);
       });    
     }
   }
   
   function fireBrushUp(canvas) {
-    canvas.freeDrawingBrush.onMouseUp(options);
+    fireBrushEvent(canvas.freeDrawingBrush, 'up', new fabric.Point());
   }
 
   var points = [
@@ -2244,7 +2249,7 @@ QUnit.module('Free Drawing', hooks => {
     }
   });
 
-    function withText(canvas) {
+  function withText(canvas) {
     canvas.add(new fabric.IText('This textbox should NOT\nclear the brush during rendering'));
     const brush = new fabric.PencilBrush(canvas);
     brush.color = 'red';
@@ -2255,17 +2260,15 @@ QUnit.module('Free Drawing', hooks => {
   tests.push({
     test: 'textbox should not clear brush',
     build: withText,
-    golden: 'withText.png',
+    name: 'withText',
     percentage: 0.02,
     width: 200,
     height: 250,
-    fabricClass: 'Canvas',
     targets: {
       top: true,
       main: false,
       mesh: true,
       result: false,
-      compare: false
     }
   });